[kubernetes] No default nodegroups in values
Since nodegroups are defined as key-value pairs in the values, there's
no easy way to remove the example md0 nodegroup that was provided there
for purposes of documentation. Commenting it out from values.yaml keeps
the documentation in place, but doesn't force the node group onto users.

Release note:
[kubernetes] Remove potentially undesirable node group from default
values of kubernetes chart.

Signed-off-by: Timofei Larkin <lllamnyp@gmail.com>