[kamaji] Pass in-tree image as migrate-image
The in-tree build of the Kamaji image lacks the appropriate ldflags,
resulting in invalid flags of the Kamaji controller manager binary. When
a migration job starts, it tries to pull an image with an explicit empty
string as a tag, which is invalid. This patch sets the in-tree image as
the image for the migration job, both working around this issue, as well
as being consistent in the image used.

Signed-off-by: Timofei Larkin <lllamnyp@gmail.com>
(cherry picked from commit fbc2c45e7f1701e568de33ca926ae370930d9d91)